Rename and refactor into partial
