+++ app/assets/stylesheets/libraries.css.scss
@@ -1,4 +1,4 @@
 @import "libraries/dragula.css.scss";
 @import "libraries/flatpickr.css.scss";
 @import "libraries/glightbox.css.scss";
-@import "libraries/papyros.css.scss";
+@import "libraries/coding_scratchpad.css.scss";
