+++ app/assets/stylesheets/libraries/coding_scratchpad.css.scss
