+++ app/javascript/packs/coding_scratchpad.js
@@ -1,0 +1,3 @@
+import { initCodingScratchpad } from "coding_scratchpad.js";
+
+window.dodona.initCodingScratchpad = initCodingScratchpad;
