+++ app/views/activities/show.html.erb
@@ -77,9 +77,6 @@ end %>
       <div class="card-title card-title-colored">
         <% if policy(@activity).submit? %>
           <div class="card-title-fab">
-            <button class="btn btn-primary btn-fab mdi mdi-pencil mdi-36" type="button" data-bs-toggle="offcanvas"
-              data-bs-target="#papyros-offcanvas" aria-controls="papyros-offcanvas" id="papyros-offcanvas-show-btn">
-            </button>
             <button class="btn btn-primary btn-fab mdi mdi-play mdi-36" id="editor-process-btn" data-user_id="<%= current_user.id %>">
             </button>
             <a class="btn btn-primary btn-fab hidden-fab" id="submission-copy-btn">
@@ -132,30 +129,7 @@ end %>
           <div class="tab-pane fade hidden" id="feedback">
             <div id="submission-wrapper"></div>
           </div>
-
-          <div class="offcanvas offcanvas-end papyros-offcanvas" tabindex="-1" id="papyros-offcanvas">
-            <div class="card-title papyros-header">
-              <div class="card-subtitle margin-bottom-11">
-                <h2 class="card-title-text">
-                  <%= activity_icon(@activity, 24) %>
-                  &nbsp;
-                  <%= t '.papyros.papyros' %>
-                </h2>
-                <div class="flex-spacer"></div>
-                <button type="button" class="btn-close btn-close-white" data-bs-dismiss="offcanvas" aria-label="<%= t '.papyros.close' %>"></button>
-              </div>
-              </div>
-            <span class="header-info-text papyros-info"><i class="mdi mdi-information mdi-18"></i> <%= t ".papyros.experimental_feature" %></span>
-            <div class="offcanvas-body">
-              <div id="papyros-editor-wrapper"></div>
-              <div id="papyros-panel-wrapper"></div>
-              <div id="papyros-output-wrapper"></div>
-              <div id="papyros-input-wrapper"></div>
-              <button class="btn btn-primary btn-fab mdi mdi-content-copy mdi-36" id="papyros-code-copy-btn" title="<%= t '.papyros.copy_code' %>">
-              </button>
-            </div>
-          </div>
-
+          <%= render partial: 'coding_scratchpad', locals: {activity: @activity} %>
         </div>
       </div>
     </div>
