+++ config/locales/views/activities/nl.yml
@@ -80,11 +80,6 @@ nl:
         next: Volgende activiteit
       mark_as_read: Markeren als gelezen
       read_at: "Gelezen op %{timestamp}"
-      papyros:
-        papyros: Papyros
-        experimental_feature: Papyros is een experimentele feature die je toelaat om je code uit te voeren in de browser en je helpt tijdens je leerproces. Suggesties en problemen kan je melden bij team Dodona.
-        copy_code: Kopieer je code naar Dodona
-        close: Sluiten
     series_activities_add_table:
       course_added_to_usable: "Deze oefening toevoegen zal deze cursus toegang geven tot alle privé oefeningen in de repository van deze oefening. Ben je zeker?"
     edit:
@@ -129,3 +124,8 @@ nl:
     series_activities_table:
       course_has_no_rights_with_fix: Deze cursus heeft niet de rechten om deze privé-oefening te gebruiken. Klik hier om deze rechten te geven.
       course_has_no_rights_with_contact: Deze cursus heeft niet de rechten om deze privé-oefening te gebruiken. Contacteer de auteur of verwijder deze oefening uit de cursus.
+    coding_scratchpad:
+      title: "Voer %{programming_language} code uit in je browser"
+      experimental_feature: Papyros is een experimentele feature die je toelaat om je code uit te voeren in de browser en je helpt tijdens je leerproces. Suggesties en problemen kan je melden bij team Dodona.
+      copy_code: Kopieer je code naar Dodona
+      close: Sluiten
